Fix x64 warnings.

+++ win10/edge/test/bound_function.cpp
@@ -182,7 +182,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -203,7 +203,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -223,7 +223,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -242,7 +242,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -260,7 +260,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -277,7 +277,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -293,7 +293,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -308,7 +308,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -322,7 +322,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             return 0;
@@ -334,7 +334,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -353,7 +353,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -371,7 +371,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -388,7 +388,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -404,7 +404,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -419,7 +419,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -433,7 +433,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -446,7 +446,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
@@ -458,7 +458,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(((jsrt::object)info.this_value()).is_external());
             void *data = ((jsrt::external_object)info.this_value()).data();
-            Assert::AreEqual(data, (void *) 0xdeadbeef);
+            Assert::AreEqual(data, (void *) (__int64) 0xdeadbeef);
             Assert::IsFalse(info.is_construct_call());
         }
 
@@ -468,7 +468,7 @@ namespace jsrtwrapperstest
             jsrt::context context = runtime.create_context();
             {
                 jsrt::context::scope scope(context);
-                jsrt::object this_value = jsrt::external_object::create((void *) 0xdeadbeef);
+                jsrt::object this_value = jsrt::external_object::create((void *) (__int64) 0xdeadbeef);
 
                 auto f8 = jsrt::bound_function<jsrt::object, double, std::wstring, double, bool, std::wstring, double, bool, std::wstring, double>::create(this_value, callback8);
                 double result = f8(L"foo", 2, true, L"bar", 5, false, L"baz", 8);
@@ -551,7 +551,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(p7, (std::wstring)L"baz");
             Assert::AreEqual(p8, 8.0);
 
-            return jsrt::external_object::create((void *) 0xdeadbeef);
+            return jsrt::external_object::create((void *) (__int64) 0xdeadbeef);
         }
 
         static jsrt::object callback7c(const jsrt::call_info &info, std::wstring p1, double p2, bool p3, std::wstring p4, double p5, bool p6, std::wstring p7)
@@ -568,7 +568,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(p6, false);
             Assert::AreEqual(p7, (std::wstring)L"baz");
 
-            return jsrt::external_object::create((void *) 0xdeadbeef);
+            return jsrt::external_object::create((void *) (__int64) 0xdeadbeef);
         }
 
         static jsrt::object callback6c(const jsrt::call_info &info, std::wstring p1, double p2, bool p3, std::wstring p4, double p5, bool p6)
@@ -584,7 +584,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(p5, 5.0);
             Assert::AreEqual(p6, false);
 
-            return jsrt::external_object::create((void *) 0xdeadbeef);
+            return jsrt::external_object::create((void *) (__int64) 0xdeadbeef);
         }
 
         static jsrt::object callback5c(const jsrt::call_info &info, std::wstring p1, double p2, bool p3, std::wstring p4, double p5)
@@ -599,7 +599,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(p4, (std::wstring)L"bar");
             Assert::AreEqual(p5, 5.0);
 
-            return jsrt::external_object::create((void *) 0xdeadbeef);
+            return jsrt::external_object::create((void *) (__int64) 0xdeadbeef);
         }
 
         static jsrt::object callback4c(const jsrt::call_info &info, std::wstring p1, double p2, bool p3, std::wstring p4)
@@ -613,7 +613,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(p3, true);
             Assert::AreEqual(p4, (std::wstring)L"bar");
 
-            return jsrt::external_object::create((void *) 0xdeadbeef);
+            return jsrt::external_object::create((void *) (__int64) 0xdeadbeef);
         }
 
         static jsrt::object callback3c(const jsrt::call_info &info, std::wstring p1, double p2, bool p3)
@@ -626,7 +626,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(p2, 2.0);
             Assert::AreEqual(p3, true);
 
-            return jsrt::external_object::create((void *) 0xdeadbeef);
+            return jsrt::external_object::create((void *) (__int64) 0xdeadbeef);
         }
 
         static jsrt::object callback2c(const jsrt::call_info &info, std::wstring p1, double p2)
@@ -638,7 +638,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(p1, (std::wstring)L"foo");
             Assert::AreEqual(p2, 2.0);
 
-            return jsrt::external_object::create((void *) 0xdeadbeef);
+            return jsrt::external_object::create((void *) (__int64) 0xdeadbeef);
         }
 
         static jsrt::object callback1c(const jsrt::call_info &info, std::wstring p1)
@@ -649,7 +649,7 @@ namespace jsrtwrapperstest
 
             Assert::AreEqual(p1, (std::wstring)L"foo");
 
-            return jsrt::external_object::create((void *) 0xdeadbeef);
+            return jsrt::external_object::create((void *) (__int64) 0xdeadbeef);
         }
 
         static jsrt::object callback0c(const jsrt::call_info &info)
@@ -658,7 +658,7 @@ namespace jsrtwrapperstest
             Assert::AreEqual(info.this_value().type(), JsObject);
             Assert::IsTrue(info.is_construct_call());
 
-            return jsrt::external_object::create((void *) 0xdeadbeef);
+            return jsrt::external_object::create((void *) (__int64) 0xdeadbeef);
         }
 
         MY_TEST_METHOD(strongly_typed_constructors, "Test strongly typed constructor functions.")
@@ -667,60 +667,60 @@ namespace jsrtwrapperstest
             jsrt::context context = runtime.create_context();
             {
                 jsrt::context::scope scope(context);
-                jsrt::object this_value = jsrt::external_object::create((void *) 0xdeadbeef);
+                jsrt::object this_value = jsrt::external_object::create((void *) (__int64) 0xdeadbeef);
                 auto f8 = jsrt::bound_function<jsrt::object, jsrt::object, std::wstring, double, bool, std::wstring, double, bool, std::wstring, double>::create(this_value, callback8c);
                 jsrt::object result = f8.construct(L"foo", 2, true, L"bar", 5, false, L"baz", 8);
                 Assert::IsTrue(result.is_valid());
                 Assert::IsTrue(result.is_external());
-                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) 0xdeadbeef);
+                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) (__int64) 0xdeadbeef);
 
                 auto f7 = jsrt::bound_function<jsrt::object, jsrt::object, std::wstring, double, bool, std::wstring, double, bool, std::wstring>::create(this_value, callback7c);
                 result = f7.construct(L"foo", 2, true, L"bar", 5, false, L"baz");
                 Assert::IsTrue(result.is_valid());
                 Assert::IsTrue(result.is_external());
-                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) 0xdeadbeef);
+                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) (__int64) 0xdeadbeef);
 
                 auto f6 = jsrt::bound_function<jsrt::object, jsrt::object, std::wstring, double, bool, std::wstring, double, bool>::create(this_value, callback6c);
                 result = f6.construct(L"foo", 2, true, L"bar", 5, false);
                 Assert::IsTrue(result.is_valid());
                 Assert::IsTrue(result.is_external());
-                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) 0xdeadbeef);
+                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) (__int64) 0xdeadbeef);
 
                 auto f5 = jsrt::bound_function<jsrt::object, jsrt::object, std::wstring, double, bool, std::wstring, double>::create(this_value, callback5c);
                 result = f5.construct(L"foo", 2, true, L"bar", 5);
                 Assert::IsTrue(result.is_valid());
                 Assert::IsTrue(result.is_external());
-                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) 0xdeadbeef);
+                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) (__int64) 0xdeadbeef);
 
                 auto f4 = jsrt::bound_function<jsrt::object, jsrt::object, std::wstring, double, bool, std::wstring>::create(this_value, callback4c);
                 result = f4.construct(L"foo", 2, true, L"bar");
                 Assert::IsTrue(result.is_valid());
                 Assert::IsTrue(result.is_external());
-                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) 0xdeadbeef);
+                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) (__int64) 0xdeadbeef);
 
                 auto f3 = jsrt::bound_function<jsrt::object, jsrt::object, std::wstring, double, bool>::create(this_value, callback3c);
                 result = f3.construct(L"foo", 2, true);
                 Assert::IsTrue(result.is_valid());
                 Assert::IsTrue(result.is_external());
-                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) 0xdeadbeef);
+                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) (__int64) 0xdeadbeef);
 
                 auto f2 = jsrt::bound_function<jsrt::object, jsrt::object, std::wstring, double>::create(this_value, callback2c);
                 result = f2.construct(L"foo", 2);
                 Assert::IsTrue(result.is_valid());
                 Assert::IsTrue(result.is_external());
-                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) 0xdeadbeef);
+                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) (__int64) 0xdeadbeef);
 
                 auto f1 = jsrt::bound_function<jsrt::object, jsrt::object, std::wstring>::create(this_value, callback1c);
                 result = f1.construct(L"foo");
                 Assert::IsTrue(result.is_valid());
                 Assert::IsTrue(result.is_external());
-                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) 0xdeadbeef);
+                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) (__int64) 0xdeadbeef);
 
                 auto f0 = jsrt::bound_function<jsrt::object, jsrt::object>::create(this_value, callback0c);
                 result = f0.construct();
                 Assert::IsTrue(result.is_valid());
                 Assert::IsTrue(result.is_external());
-                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) 0xdeadbeef);
+                Assert::AreEqual(((jsrt::external_object)result).data(), (void *) (__int64) 0xdeadbeef);
             }
             runtime.dispose();
         }
@@ -740,7 +740,7 @@ namespace jsrtwrapperstest
             jsrt::context context = runtime.create_context();
             {
                 jsrt::context::scope scope(context);
-                jsrt::object this_value = jsrt::external_object::create((void *) 0xdeadbeef);
+                jsrt::object this_value = jsrt::external_object::create((void *) (__int64) 0xdeadbeef);
                 jsrt::object proto = jsrt::object::create();
                 auto f0 = jsrt::bound_function<jsrt::object, jsrt::object>::create(this_value, callback0cp);
                 f0.set_constructor_prototype(proto);
